requirements.txt
Django==2.2.4
psycopg2-binary
gunicorn
PyMySQL
mysqlclient
djangorestframework
django-filter